Fix race condition in LabelMarkupText
In a parallel pipeline script multiple logger write operations can occur
against the same TaskListener logger. LabelMarkupText nulls out the
annotation field after write which could cause a NPE in a multi threaded
pipeline because the same LabelMarkupText instance is used for all
write operations.

This change doesn't address the larger thread safety issue with
PipelineConsoleDecoder and LabelMarkupText but should prevent
a possible NPE in case the the class fields are nulled out between
the isEmpty check and the value reference during write.